refactor(plugin): extract schema definitions to plugin/schema.rs (#587)
Move builtin_command_infos() and all_known_commands() from plugin/mod.rs
to a dedicated plugin/schema.rs module. Tests move with their functions.

plugin/mod.rs now contains only PluginRegistry, RESERVED_COMMANDS, and
re-exports (~160 prod lines). schema.rs holds all built-in command
schema declarations (~203 prod lines). Both well under 500-line target.

Closes #587
